add function to cleanup api state
